requirements.txt
numpy
matplotlib
torch
torchvision
timm
tqdm
sklearn
scikit-image
pandas
Pillow
openpyxl

#NLP packages
transformers

efficientnet_pytorch